requirements.txt
asgiref==3.4.1
Django==3.2.7
django-redis==5.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
pytz==2021.3
redis==3.5.3
sqlparse==0.4.2
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
